refactoring definitions
