+++ definitions/hcl/intermediate.go
@@ -1,4 +1,4 @@
-package intermediate
+package hcl
 
 import (
 	"github.com/hashicorp/hcl/v2"
